chore: move videoconf context to ui-contexts
