+++ apps/meteor/client/uikit/hooks/useMessageBlockContextValue.ts
@@ -1,17 +1,17 @@
 import type { IRoom, IMessage } from '@rocket.chat/core-typings';
 import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import type { UiKitContext } from '@rocket.chat/fuselage-ui-kit';
-import type { ContextType } from 'react';
-
-import { useUiKitActionManager } from './useUiKitActionManager';
 import {
 	useVideoConfDispatchOutgoing,
 	useVideoConfIsCalling,
 	useVideoConfIsRinging,
 	useVideoConfJoinCall,
-	useVideoConfManager,
+	useVideoConfLoadCapabilities,
 	useVideoConfSetPreferences,
-} from '../../contexts/VideoConfContext';
+} from '@rocket.chat/ui-contexts';
+import type { ContextType } from 'react';
+
+import { useUiKitActionManager } from './useUiKitActionManager';
 import { useVideoConfWarning } from '../../views/room/contextualBar/VideoConference/hooks/useVideoConfWarning';
 
 export const useMessageBlockContextValue = (rid: IRoom['_id'], mid: IMessage['_id']): ContextType<typeof UiKitContext> => {
@@ -21,8 +21,7 @@ export const useMessageBlockContextValue = (rid: IRoom['_id'], mid: IMessage['_i
 	const isRinging = useVideoConfIsRinging();
 	const dispatchWarning = useVideoConfWarning();
 	const dispatchPopup = useVideoConfDispatchOutgoing();
-
-	const videoConfManager = useVideoConfManager();
+	const loadVideoConfCapabilities = useVideoConfLoadCapabilities();
 
 	const handleOpenVideoConf = useEffectEvent(async (rid: IRoom['_id']) => {
 		if (isCalling || isRinging) {
@@ -30,7 +29,7 @@ export const useMessageBlockContextValue = (rid: IRoom['_id'], mid: IMessage['_i
 		}
 
 		try {
-			await videoConfManager?.loadCapabilities();
+			await loadVideoConfCapabilities();
 			dispatchPopup({ rid });
 		} catch (error: any) {
 			dispatchWarning(error.error);
